requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
Deprecated==1.2.15
Flask==3.1.0
fpdf==1.7.2
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
keyboard==0.13.5
lxml==5.3.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
oauthlib==3.2.2
packaging==24.2
pikepdf==9.4.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyKCS11==1.5.17
PyMsgBox==1.0.9
pyparsing==3.2.1
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==308
pywinauto==0.6.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.16.0
